+++ modules/mapbox/src/mapbox-layer-group.ts
@@ -2,40 +2,39 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-import {drawLayer} from './deck-utils';
+import {drawLayerGroup} from './deck-utils';
 import type {Map, CustomLayerInterface} from './types';
-import type {Deck, Layer} from '@deck.gl/core';
+import {assert, type Deck} from '@deck.gl/core';
 
 type MapWithDeck = Map & {__deck: Deck};
 
-export type MapboxLayerProps<LayerT extends Layer> = Partial<LayerT['props']> & {
+export type MapboxLayerGroupProps = {
   id: string;
   renderingMode?: '2d' | '3d';
   /* Mapbox v3 Standard style */
   slot?: 'bottom' | 'middle' | 'top';
+  beforeId?: string;
 };
 
-export default class MapboxLayer<LayerT extends Layer> implements CustomLayerInterface {
+export default class MapboxLayerGroup implements CustomLayerInterface {
   id: string;
   type: 'custom';
   renderingMode: '2d' | '3d';
   /* Mapbox v3 Standard style */
   slot?: 'bottom' | 'middle' | 'top';
+  beforeId?: string;
   map: MapWithDeck | null;
-  props: MapboxLayerProps<LayerT>;
 
   /* eslint-disable no-this-before-super */
-  constructor(props: MapboxLayerProps<LayerT>) {
-    if (!props.id) {
-      throw new Error('Layer must have an unique id');
-    }
+  constructor(props: MapboxLayerGroupProps) {
+    assert(props.id, 'id is required');
 
     this.id = props.id;
     this.type = 'custom';
     this.renderingMode = props.renderingMode || '3d';
     this.slot = props.slot;
+    this.beforeId = props.beforeId;
     this.map = null;
-    this.props = props;
   }
 
   /* Mapbox custom layer methods */
@@ -44,18 +43,9 @@ export default class MapboxLayer<LayerT extends Layer> implements CustomLayerInt
     this.map = map;
   }
 
-  onRemove(): void {
-    this.map = null;
-  }
-
-  setProps(props: MapboxLayerProps<LayerT>) {
-    // id cannot be changed
-    Object.assign(this.props, props, {id: this.id});
-  }
-
   render(gl, renderParameters) {
     if (!this.map) return;
 
-    drawLayer(this.map.__deck, this.map, this, renderParameters);
+    drawLayerGroup(this.map.__deck, this.map, this, renderParameters);
   }
 }