+++ modules/mapbox/src/mapbox-overlay.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-import {Deck, assert} from '@deck.gl/core';
+import {Deck, assert, log} from '@deck.gl/core';
 import {
   getViewState,
   getDefaultView,
@@ -14,10 +14,10 @@ import {
 
 import type {Map, IControl, MapMouseEvent, ControlPosition} from './types';
 import type {MjolnirGestureEvent, MjolnirPointerEvent} from 'mjolnir.js';
-import type {DeckProps} from '@deck.gl/core';
-import {log} from '@deck.gl/core';
+import type {DeckProps, LayersList} from '@deck.gl/core';
 
 import {resolveLayers} from './resolve-layers';
+import {resolveLayerGroups} from './resolve-layer-groups';
 
 export type MapboxOverlayProps = Omit<
   DeckProps,
@@ -31,7 +31,16 @@ export type MapboxOverlayProps = Omit<
   | 'initialViewState'
   | 'controller'
 > & {
+  /**
+   * deck.gl layers are inserted into mapbox-gl's layer stack, and share the same WebGL2RenderingContext as the base map.
+   */
   interleaved?: boolean;
+
+  /**
+   * (experimental) render deck.gl layers in batches grouped by `beforeId` or `slot` to enable cross-layer extension handling.
+   * @default false
+   */
+  _renderLayersInGroups?: boolean;
 };
 
 /**
@@ -44,11 +53,13 @@ export default class MapboxOverlay implements IControl {
   private _map?: Map;
   private _container?: HTMLDivElement;
   private _interleaved: boolean;
+  private _renderLayersInGroups: boolean;
   private _lastMouseDownPoint?: {x: number; y: number; clientX: number; clientY: number};
 
   constructor(props: MapboxOverlayProps) {
     const {interleaved = false} = props;
     this._interleaved = interleaved;
+    this._renderLayersInGroups = props._renderLayersInGroups || false;
     this._props = this.filterProps(props);
   }
 
@@ -65,7 +76,7 @@ export default class MapboxOverlay implements IControl {
   /** Update (partial) props of the underlying Deck instance. */
   setProps(props: MapboxOverlayProps): void {
     if (this._interleaved && props.layers) {
-      resolveLayers(this._map, this._deck, this._props.layers, props.layers);
+      this._resolveLayers(this._map, this._deck, this._props.layers, props.layers);
     }
 
     Object.assign(this._props, this.filterProps(props));
@@ -126,7 +137,7 @@ export default class MapboxOverlay implements IControl {
 
   private _onAddInterleaved(map: Map): HTMLDivElement {
     // @ts-ignore non-public map property
-    const gl = map.painter.context.gl;
+    const gl: WebGL2RenderingContext = map.painter.context.gl;
     if (gl instanceof WebGLRenderingContext) {
       log.warn(
         'Incompatible basemap library. See: https://deck.gl/docs/api-reference/mapbox/overview#compatibility'
@@ -134,7 +145,6 @@ export default class MapboxOverlay implements IControl {
     }
     this._deck = getDeckInstance({
       map,
-      gl,
       deck: new Deck({
         ...this._props,
         gl,
@@ -143,9 +153,22 @@ export default class MapboxOverlay implements IControl {
     });
 
     map.on('styledata', this._handleStyleChange);
-    resolveLayers(map, this._deck, [], this._props.layers);
+    this._resolveLayers(map, this._deck, [], this._props.layers);
 
     return document.createElement('div');
+  }
+
+  private _resolveLayers(
+    map: Map | undefined,
+    deck: Deck | undefined,
+    prevLayers: LayersList | undefined,
+    newLayers: LayersList | undefined
+  ): void {
+    if (this._renderLayersInGroups) {
+      resolveLayerGroups(map, prevLayers, newLayers);
+    } else {
+      resolveLayers(map, deck, prevLayers, newLayers);
+    }
   }
 
   /** Called when the control is removed from a map */
@@ -181,7 +204,7 @@ export default class MapboxOverlay implements IControl {
 
   private _onRemoveInterleaved(map: Map): void {
     map.off('styledata', this._handleStyleChange);
-    resolveLayers(map, this._deck, this._props.layers, []);
+    this._resolveLayers(map, this._deck, this._props.layers, []);
     removeDeckInstance(map);
   }
 
@@ -225,7 +248,7 @@ export default class MapboxOverlay implements IControl {
   }
 
   private _handleStyleChange = () => {
-    resolveLayers(this._map, this._deck, this._props.layers, this._props.layers);
+    this._resolveLayers(this._map, this._deck, this._props.layers, this._props.layers);
     if (!this._map) return;
 
     // getProjection() returns undefined before style is loaded