+++ modules/aggregation-layers/src/heatmap-layer/heatmap-layer.ts
@@ -329,7 +329,8 @@ export default class HeatmapLayer<
   _getAttributeManager() {
     return new AttributeManager(this.context.device, {
       id: this.props.id,
-      stats: this.context.stats
+      stats: this.context.stats,
+      memory: this.props.memory
     });
   }